requirements.txt
zulip==0.9.0
google-cloud-aiplatform>=1.38.0
requests>=2.28.0
python-dotenv>=1.0.0
click>=8.0.0